cpufreq: Allow user min/max changes independent of policy min/max
When user min/max values are written to scaling_min/max_freq sysfs files,
the current code applies all the limits imposed by various ADJUST and
INCOMPATIBLE notifiers handlers before storing the user min/max in the
policy.

This behavior makes it impossible for a user/userspace process to reliably
set the user min/max limits outside the range of whatever instantaneous
limits that are being imposed by the ADJUST/INCOMPATIBLE notifiers.  Since
these sysfs nodes are typically meant to set limits that need to be
enforced for the foreseeable future, this is not a very usable or user
friendly behavior.

So, change the behavior to only apply the CPU limits since that doesn't
change at runtime.  This patch continues to enforce the requirement that
user min has to be less than or equal to user max.

Change-Id: I7143bc7ef1811defd48ba91619d96896a3dd36e7
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
[junjiew@codeaurora.org: conflict resolution]
Signed-off-by: Junjie Wu <junjiew@codeaurora.org>