cpufreq: gabriel*: calculate load before freq change
The update to cpu load for cpufreq rate changes was happening after the rate
change instead of before. This switches it to update the cpu load before the
cpufreq rate change so the old frequency is used for calculating the speed
adjusted cpu load. The old frequency was used for that active time, so it should
be used for calculating the speed adjusted cpu load.

original-commit:
https://github.com/nikhil18/Lightning_kernel_oneplus5/commit/5998056c30fe1e42b109657fe0a42d867c0b657a

Signed-off-by: mostafa-z <mostafazarghami@gmail.com>